refactor: make QueryBlocklistRule an interface for extensibility
